This commit re-implements the IScalarUnmarshallingContext from java into C#, with the single method UriContext() that returns a Uri object.
All scalarTypes will now receive an object that implements the interface.
UriType uses this uriContext when available to make valid Uris out of relative Uri strings that it is deserializing.

TranslationScope.deserialize takes an optional argument uriContext, that is used by the JSON and XML parsers.

deserializeString for JSON also takes an optional string argument of the jsonText. Since the JsonReader has another constructor that accepts string directly, it uses the jsonText when available, in which case the stream parameter can be null.

These changes shouldn't affect anything outside of fundamental.NET